+++ db/migrate/20221103083443_add_default_gravity_to_files.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddDefaultGravityToFiles < ActiveRecord::Migration[7.0]
+  def change
+    add_column :folio_files, :default_gravity, :string
+  end
+end